Added rudimentary history generation.
